Major changes this time folks.
1) Completely tear apart parse_input.py. Change the main function in this
module to init, which returns a list of statements. Remove quite a bit
to other modules, so that this one only parses the input.

	i) Reduce bulk of the statement class so that it only contains
	the parsing information, and not algorithmic.

	ii) Create a new algorithmic module called alg_line_context,
	where the algorithmic parsing has been transferred to, by
	deriving a special class from statement just for this
	module.

2) Create the settings module, which ,for now,  contains just the list of
algorithms to use.

3) Implement the run_algorithms module, which runs all the algorithms
specified in the settings module, and does a multiplication of the values
obtained.

4) Implement the alg_line_context module. Use this module to show
how to create an algorithm module. Basically the run_alg function should
return a dictionary of user and pseudo-probabilities for a given
list of statements and the index of the unknown one.

5) The complete description of how to write an algorithm module and
how to use the run_algorithms module to run it, as well as recommended
design for easy debugging, will be added in a subsequent commit.

Good work for a day!